requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
environs[django]==9.3.2
Pillow==10.1.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.26
